Create FundMe.sol
Contract for accepting ether from people and being able to withdraw it. Only the contract owner is able to call the withdraw function and we also have a receive and fallback function to call the fund function if someone mistakenly sends ether to the contract without calling the fund function or sends a calldata which does not represent any function in the contract.
Also made use of chainlink oracle to make the minimum funding amount in ether equivalent to the $50.